cloud_run_service/requirements.txt
flask
gunicorn
google-cloud-storage
google-cloud-firestore
transformers
torch
PyMuPDF
